tests/requirements.txt
asyncio==3.4.3
async-timeout==5.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
redis==5.2.0
pytest-asyncio==0.24.0
aio-pika==9.5.3
backoff==2.2.1
